backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
email-validator>=2.0.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
redis>=5.0.1
httpx>=0.25.1
google-auth>=2.23.0
requests>=2.31.0
apscheduler>=3.10.0
slowapi>=0.1.9
